Add support for Long values (large numbers) instead of int

+++ src/main/kotlin/land/tbp/y2019/day_5_sunny_with_a_chance_of_asteroids/Problem1.kt
@@ -1,9 +1,7 @@
 package land.tbp.y2019.day_5_sunny_with_a_chance_of_asteroids
 
-import land.tbp.y2019.intcode.computer.CPU
 import land.tbp.y2019.intcode.computer.Computer
 import land.tbp.y2019.intcode.computer.Memory
-import land.tbp.y2019.intcode.computer.instructions.*
 import loadResourceFile
 import stringToInt
 
@@ -11,7 +9,7 @@ fun main() {
     val text = loadResourceFile("./land/tbp/y2019/day_5_sunny_with_a_chance_of_asteroids/in1.txt")
     val ints = stringToInt(text)
 
-    val outputs = mutableListOf<Int>()
+    val outputs = mutableListOf<Long>()
     Computer(
             Memory(ints),
             mutableListOf(1),
